+++ src/core/utils/index.js
@@ -1,15 +1,17 @@
 import * as base64 from './base64.js';
 import copyText from './copy-text.js';
+import defined from './defined.js';
 import * as dom from './dom.js';
 import * as html from './html.js';
-import defined from './defined.js';
+import * as layout from './layout.js';
 import safeFilterRx from './safe-filter-rx.js';
 
 export default {
     base64,
     copyText,
+    defined,
     ...dom,
     ...html,
-    defined,
+    ...layout,
     safeFilterRx
 };